New mapping file...
NEW AIRR Version: Adds a new column airr_is_array and ir_repository_is_array. Adds onotology for disease_diagnosis. Also fixes the gateway (ir_short and ir_full) names for rearrangements. Changed the name of the airr_full column to airr_name, which matches the field x-airr:name in the AIRR spec. This means adding a new row (Row 3), which should get removed on export. Changed the name of the mapping column for airr_desc to be airr_description - which is derived from the AIRR spec.